+++ src/main/java/be/baur/sds/serialization/Parser.java
@@ -3,8 +3,8 @@ package be.baur.sds.serialization;
 import java.io.IOException;
 import java.io.Reader;
 import java.io.StringReader;
+import java.text.ParseException;
 
-import be.baur.sda.serialization.SyntaxException;
 import be.baur.sds.Schema;
 
 
@@ -20,7 +20,7 @@ public interface Parser {
 	 * Parses a character <code>input</code> stream and creates a {@link Schema}.
 	 * @return Schema - the schema node.
 	 */
-	Schema parse(Reader input) throws IOException, SyntaxException, SchemaException;
+	Schema parse(Reader input) throws IOException, ParseException, SchemaException;
 
 
 	/**
@@ -29,7 +29,7 @@ public interface Parser {
 	 * by another type of parser. The default implementation serializes the schema
 	 * to SDS format, which is then parsed to reveal any issues.
 	 */
-	default void verify(Schema schema) throws SyntaxException, IOException, SchemaException {
+	default void verify(Schema schema) throws IOException, ParseException, SchemaException {
 		parse(new StringReader(schema.toString()));
 	}
 }