+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@ validation of SDA documents, syntax version 2. !!! Work In Proress !!!
 - `Added` NodeType to replace SimpleType and ComplexType.
 - `Changed` ComponentType to an abstract class (was an interface).
 - `Changed` AbstractGroup to extend ComponentType (was ComplexType).
+- `Changed` Parser.parse() to throw ParseException (was SyntaxException).
 - `Renamed` AbstractGroup to ModelGroup, Group to SequenceGroup.
 - `Removed` SimpleType and ComplexType (replaced with NodeType).